requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
anyio==4.8.0
astroid==3.3.8
blinker==1.9.0
cachelib==0.9.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
config==0.5.1
dashboard==0.0.6
databases==0.9.0
dill==0.3.9
distlib==0.3.9
Faker==35.2.0
filelock==3.17.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-SQLAlchemy==3.1.1
h11==0.14.0
identify==2.6.6
idna==3.10
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kafka-python==2.0.3
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
orm==0.2.0.dev1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
psycopg2-binary==2.9.10
pylint==3.3.4
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomlkit==0.13.2
typesystem==0.3.0.dev0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.1
Werkzeug==3.1.3
